refactor: Menu/Product 관련 도메인/테스트 코드 리팩토링, 미사용 코드 삭제
- JPA Repository로 DB 로직을 대체함에 따라 dao 코드 삭제
- Menu 및 Product의 도메인 코드 리팩토링, 패키지 이동
- Menu 및 Product의 컨트롤러, 서비스 코드 리팩토링, 패키지 이동
- MenuProduct의 1급 컬렉션에 대한 테스트 코드 패키지 이동
- 사용하지 않는 테스트의 Creator 팩토리 클래스 삭제
