+++ src/main/java/kitchenpos/domain/menu/MenuProductRepository.java
@@ -1,11 +1,9 @@
 package kitchenpos.domain.menu;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
-@Repository
 public interface MenuProductRepository extends JpaRepository<MenuProduct, Long> {
     public List<MenuProduct> findAllByMenu(Menu menu);
 }
