+++ src/main/java/kitchenpos/domain/menu/MenuRepository.java
@@ -1,8 +1,6 @@
 package kitchenpos.domain.menu;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
-@Repository
 public interface MenuRepository extends JpaRepository<Menu, Long> {
 }
