+++ src/main/java/kitchenpos/dto/menu/MenuGroupResponse.java
@@ -1,20 +1,19 @@
-package kitchenpos.dto;
+package kitchenpos.dto.menu;
 
 import kitchenpos.domain.menu.MenuGroup;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import javax.validation.constraints.NotBlank;
-
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class MenuGroupCreateRequest {
-    @NotBlank
+public class MenuGroupResponse {
+    private Long id;
     private String name;
 
-    public MenuGroup toMenuGroup() {
-        return new MenuGroup(name);
+    public MenuGroupResponse(MenuGroup menuGroup) {
+        this.id = menuGroup.getId();
+        this.name = menuGroup.getName();
     }
 }
