+++ src/main/java/kitchenpos/dto/menu/ProductResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.menu;
 
 import kitchenpos.domain.menu.Product;
 import lombok.AllArgsConstructor;
