+++ src/main/java/kitchenpos/ui/MenuRestController.java
@@ -1,8 +1,9 @@
 package kitchenpos.ui;
 
 import kitchenpos.application.MenuService;
-import kitchenpos.dto.MenuCreateRequest;
-import kitchenpos.dto.MenuResponse;
+import kitchenpos.dto.menu.MenuCreateRequest;
+import kitchenpos.dto.menu.MenuResponse;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -14,12 +15,9 @@ import java.net.URI;
 import java.util.List;
 
 @RestController
+@RequiredArgsConstructor
 public class MenuRestController {
     private final MenuService menuService;
-
-    public MenuRestController(final MenuService menuService) {
-        this.menuService = menuService;
-    }
 
     @PostMapping("/api/menus")
     public ResponseEntity<MenuResponse> create(@RequestBody @Valid final MenuCreateRequest request) {
