+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -1,7 +1,7 @@
 package kitchenpos.application;
 
 import kitchenpos.domain.menu.*;
-import kitchenpos.dto.*;
+import kitchenpos.dto.menu.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
