+++ src/test/java/kitchenpos/dto/menu/ProductQuantityRequestsTest.java
@@ -1,7 +1,5 @@
-package kitchenpos.domain.menu;
+package kitchenpos.dto.menu;
 
-import kitchenpos.dto.ProductQuantityRequest;
-import kitchenpos.dto.ProductQuantityRequests;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
