+++ src/test/java/kitchenpos/ui/MenuRestControllerTest.java
@@ -2,7 +2,7 @@ package kitchenpos.ui;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import kitchenpos.application.MenuService;
-import kitchenpos.dto.*;
+import kitchenpos.dto.menu.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
